# KebabBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the project.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get all user's comments",
                "description": "Get all comments belonging to the authenticated user.",
                "operationId": "57468ee662aad477bea96c8d0a63e494",
                "responses": {
                    "200": {
                        "description": "List of comments"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/kebabs/{kebab}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get all comments for a specific kebab",
                "description": "Get all comments for a specific kebab.",
                "operationId": "e25417f0243f90f94546a0bff63bf7b2",
                "parameters": [
                    {
                        "name": "kebab",
                        "in": "path",
                        "description": "ID of the kebab",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of comments"
                    },
                    "404": {
                        "description": "Kebab not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add a new comment to a kebab",
                "description": "Add a new comment to a specific kebab.",
                "operationId": "ab65efa23dedbf890921ad60a457866e",
                "parameters": [
                    {
                        "name": "kebab",
                        "in": "path",
                        "description": "ID of the kebab",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Great kebab!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment added successfully"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/comments/{comment}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Edit user's comment",
                "description": "Edit a comment belonging to the authenticated user.",
                "operationId": "706025f7ca2d2b7bf54602535ccd20b1",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Updated comment text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Remove user's comment",
                "description": "Remove a comment belonging to the authenticated user.",
                "operationId": "82a74e5a6dacd6f3a96ef638568fbafe",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/delete-comment/{comment}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Admin can remove any comment",
                "description": "Admin can remove any comment.",
                "operationId": "75bb4e82946cf71881c73c9288f35357",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": [],
                        "admin": []
                    }
                ]
            }
        },
...
    },
...
}